Revert "Follow dependencies for model npm packages"
This reverts commit 7b0b574e067ce830dc33365c1ebb35e6e7832b47.
