+++ packages/composer-common/test/data/zip/test-npm-archive/package.json
@@ -3,7 +3,6 @@
   "version": "0.0.1",
   "description": "A test business network using npm model dependencies.",
   "dependencies": {
-    "animaltracking-model": "0.0.7",
-    "composer.business": "latest"
+    "animaltracking-model": "0.0.7"
   }
 }
